Using a library of groups

A library is represented by a handle, which is the first argument of the functions below. Which groups a library provides, and whether their number and their identification are available, depends on the library.

Expression	Meaning

L[k, i]	the group with identifier (k, i)

find(L, filters...)	a lazy selection S of groups

for G in S, collect(S), first(S), isempty(S)	the groups in S

length(S)	the number of groups in S

keys(S)	the identifiers of the groups in S

identify(L, G)	the identifier of the group G

GroupLibrary — Type

GroupLibrary{T}

Abstract supertype of the handles for libraries of groups; T is the type of the groups a handle returns.

A handle L is created by one of small_groups_library, transitive_groups_library, primitive_groups_library, perfect_groups_library, groups_with_class_number_library. Each group in a library has an identifier. For the libraries above it is a pair (k, i) where k is the value of the primary key of the library (the order, the degree, ...) and i numbers the groups with that value.

L[k, i] and L[(k, i)] return the group with identifier (k, i).

find(L::GroupLibrary, filters...) selects the groups with given properties.

identify returns the identifier of a given group.

has_groups, has_number_of_groups and has_identification tell which part of a library is available.

Examples

julia> L = transitive_groups_library()
Library of transitive groups

julia> G = L[4, 3]
Permutation group of degree 4

julia> identify(L, G)
(4, 3)

find — Method

find(L::GroupLibrary, filters...)

Return a lazy selection of the groups in L that satisfy all of filters. Each filter has one of the following forms.

func => value selects groups for which func returns value

func => list selects groups for which func returns an element of list

func selects groups for which func returns true

!func selects groups for which func returns false

At least one filter must restrict the primary key of L. A leading integer or vector of integers abbreviates such a filter, so find(L, 16) means find(L, order => 16) for a library whose primary key is the order. The documentation of each library lists the functions it supports.

Creating the selection S computes nothing. Afterwards

iterating S yields the groups, by increasing primary key and number; they are constructed one at a time,

collect(S) returns them as a vector,

first(S) returns the first group, isempty(S) tells whether there is one,

length(S) returns the number of groups,

keys(S) iterates over the identifiers of the groups.

length and keys construct groups only if the library cannot decide a filter from precomputed data.

Examples

julia> L = small_groups_library();

julia> S = find(L, order => 16, !is_abelian)
Selection of small groups: order => 16, !is_abelian

julia> length(S)
9

julia> first(S)
Pc group of order 16

julia> collect(keys(S))
9-element Vector{Tuple{ZZRingElem, ZZRingElem}}:
 (16, 3)
 (16, 4)
 (16, 6)
 (16, 7)
 (16, 8)
 (16, 9)
 (16, 11)
 (16, 12)
 (16, 13)

julia> [describe(G) for G in find(L, 8)]
5-element Vector{String}:
 "C8"
 "C4 x C2"
 "D8"
 "Q8"
 "C2 x C2 x C2"

julia> length(find(L, order => 512, is_abelian))
30

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

identify — Function

identify(L::GroupLibrary, G)

Return the identifier (k, i) of the group G in the library L, that is, L[k, i] and G are equivalent in the sense of L (isomorphic, or permutation isomorphic). An exception is thrown if G does not belong to L or if has_identification(L, k) is false, where k is the value of the primary key of L for G.

Examples

julia> identify(small_groups_library(), symmetric_group(4))
(24, 12)

julia> identify(primitive_groups_library(), symmetric_group(4))
(4, 2)

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

has_groups — Function

has_groups(L::GroupLibrary, k::IntegerUnion)

Return whether the groups in L with primary key k are available.

Examples

julia> has_groups(small_groups_library(), 768)
true

julia> has_groups(small_groups_library(), 1024)
false

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

has_number_of_groups — Function

has_number_of_groups(L::GroupLibrary, k::IntegerUnion)

Return whether the number of groups in L with primary key k is available, that is, whether length(find(L, k)) works.

Examples

julia> has_number_of_groups(small_groups_library(), 1024)
true

julia> length(find(small_groups_library(), 1024))
49487367289

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

has_identification — Function

has_identification(L::GroupLibrary, k::IntegerUnion)

Return whether identify is available for the groups in L with primary key k.

Examples

julia> has_identification(small_groups_library(), 256)
true

julia> has_identification(small_groups_library(), 512)
false

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

Available libraries

small_groups_library — Function

small_groups_library(::Type{T} = GAPGroup)

Return a handle for the library of groups of small order, see Oscar.GroupLibrary for what one can do with it.

The primary key is the order; the group L[n, i] is the i-th group of order n, up to isomorphism. The groups are provided by the GAP packages SmallGrp [SmallGrp], SOTGrps [SOTGrps] and SglPPow [SglPPow].

A group is returned as PcGroup if it is solvable and as PermGroup otherwise; if T is given, each group is converted to type T.

find(L::GroupLibrary, filters...) supports the functions exponent, is_abelian, is_almost_simple, is_cyclic, is_nilpotent, is_perfect, is_quasisimple, is_simple, is_sporadic_simple, is_solvable, is_supersolvable, number_of_conjugacy_classes, order. For most orders the library knows is_abelian, is_nilpotent, is_solvable and is_supersolvable of its groups without constructing them.

Examples

julia> L = small_groups_library()
Library of small groups

julia> L[60, 5]
Permutation group of degree 5 and order 60

julia> length(find(L, order => 512, !is_abelian))
10494183

julia> small_groups_library(PermGroup)[8, 3]
Permutation group of degree 4 and order 8

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

transitive_groups_library — Function

transitive_groups_library()

Return a handle for the library of transitive permutation groups, see Oscar.GroupLibrary for what one can do with it.

The primary key is the degree; the group L[d, i] is the i-th transitive group on d points, up to permutation isomorphism. The groups are provided by the GAP package TransGrp [TransGrp], for degrees up to 48; the data for the degrees 32 and 48 are not installed with OSCAR. The numbering for degree up to 15 is that of [CHM98].

A permutation group is regarded as a group on its moved points: identify(L, G) returns (d, i) where d is number_of_moved_points(G), which can be smaller than degree(G). The two agree for the groups in the library, and find(L::GroupLibrary, filters...) accepts number_of_moved_points in place of degree.

find(L::GroupLibrary, filters...) supports the functions degree, exponent, is_abelian, is_almost_simple, is_cyclic, is_nilpotent, is_perfect, is_primitive, is_quasisimple, is_simple, is_sporadic_simple, is_solvable, is_supersolvable, is_transitive, number_of_conjugacy_classes, number_of_moved_points, order, transitivity. A selection with filters besides the degree constructs the matching groups of one degree together.

Examples

julia> L = transitive_groups_library()
Library of transitive groups

julia> L[5, 4]
Alternating group of degree 5

julia> collect(find(L, degree => 3:5, is_abelian))
4-element Vector{PermGroup}:
 Alternating group of degree 3
 Permutation group of degree 4
 Permutation group of degree 4
 Permutation group of degree 5

julia> length(find(L, 30))
5712

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

primitive_groups_library — Function

primitive_groups_library()

Return a handle for the library of primitive permutation groups, see Oscar.GroupLibrary for what one can do with it.

The primary key is the degree; the group L[d, i] is the i-th primitive group on d points, up to permutation isomorphism. The groups are provided by the GAP package PrimGrp [PrimGrp].

A permutation group is regarded as a group on its moved points: identify(L, G) returns (d, i) where d is number_of_moved_points(G), which can be smaller than degree(G). The two agree for the groups in the library, and find(L::GroupLibrary, filters...) accepts number_of_moved_points in place of degree.

find(L::GroupLibrary, filters...) supports the same functions as for transitive_groups_library.

Examples

julia> L = primitive_groups_library()
Library of primitive groups

julia> L[10, 1]
Permutation group of degree 10 and order 60

julia> G = stabilizer(symmetric_group(5), 1)[1];

julia> degree(G), identify(L, G)
(5, (4, 2))

julia> S = find(L, degree => 10, !is_solvable, order => 1:1000)
Selection of primitive groups: degree => 10, !is_solvable, order => 1:1000

julia> collect(keys(S))
6-element Vector{Tuple{Int64, Int64}}:
 (10, 1)
 (10, 2)
 (10, 3)
 (10, 4)
 (10, 5)
 (10, 6)

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

perfect_groups_library — Function

perfect_groups_library(::Type{T} = PermGroup)

Return a handle for the library of perfect groups, see Oscar.GroupLibrary for what one can do with it.

The primary key is the order; the group L[n, i] is the i-th perfect group of order n, up to isomorphism. The library contains the perfect groups of order less than $2 \cdot 10^6$, see [HP89] and [Hul22]; they are provided by the GAP package PerfGrp [PerfGrp].

The groups have type T, which can be PermGroup or FPGroup.

find(L::GroupLibrary, filters...) supports every function that is defined for groups of type T. The library stores no properties of its groups, hence a filter besides the order is evaluated on each group of the given orders.

Examples

julia> L = perfect_groups_library()
Library of perfect groups

julia> L[120, 1]
Permutation group of degree 24 and order 120

julia> perfect_groups_library(FPGroup)[120, 1]
Finitely presented group of order 120

julia> collect(find(L, order => 1:200, !is_simple))
2-element Vector{PermGroup}:
 Permutation group of degree 1 and order 1
 Permutation group of degree 24 and order 120

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.

groups_with_class_number_library — Function

groups_with_class_number_library(::Type{T} = GAPGroup)

Return a handle for the library of groups with few conjugacy classes, see Oscar.GroupLibrary for what one can do with it.

The primary key is number_of_conjugacy_classes; the group L[k, i] is the i-th group with k conjugacy classes, up to isomorphism. The groups are provided by the GAP package SmallClassNr [SmallClassNr], for k up to 14.

A group is returned as PcGroup if it is solvable and as PermGroup otherwise; if T is given, each group is converted to type T.

find(L::GroupLibrary, filters...) supports the same functions as for small_groups_library. The library knows the order of its groups without constructing them.

Examples

julia> L = groups_with_class_number_library()
Library of groups with given class number

julia> L[5, 6]
Pc group of order 21

julia> length(find(L, 14))
93

julia> [describe(G) for G in find(L, 6, order => 1:20)]
5-element Vector{String}:
 "C6"
 "C3 : C4"
 "D12"
 "D18"
 "(C3 x C3) : C2"

Experimental

This function is part of the experimental code in Oscar. Please read here for more details.